+++ sim.h
@@ -16,20 +16,20 @@ void
 GetArgs(int argc, char **argv, char **charPart, int argChar, int *useOwnScheduler, int *numProcessors, char **fileName);
 
 void HandleOneCPU(input_node_ptr input_list_head, int numMainProcess, input_node_ptr input_process_ptr,
-                  unsigned int *globalTimer, unsigned int *totalDeltaTime, int *numRunnedProcesses,
+                  unsigned long *globalTimer, unsigned int *totalDeltaTime, int *numRunnedProcesses,
                   double *maxTimeOverhead, double *totalTimeOverhead);
 
 void HandleMultiCPU(input_node_ptr input_list_head, int numMainProcess, int numCPU, HashTable *parallelProcTable,
-                    input_node_ptr input_process_ptr, unsigned int *globalTimer, unsigned int *totalDeltaTime,
+                    input_node_ptr input_process_ptr, unsigned long *globalTimer, unsigned int *totalDeltaTime,
                     int *numRunnedProcesses, double *maxTimeOverhead, double *totalTimeOverhead);
 
 
 process_t *
-HandleArrivingProcOnOneCpu(input_node_ptr input_list_head, input_node_ptr input_process_ptr, unsigned int *globalTimer,
+HandleArrivingProcOnOneCpu(input_node_ptr input_list_head, input_node_ptr input_process_ptr, unsigned long *globalTimer,
                            unsigned int *totalDeltaTime, int *numRunnedProcesses, double *maxTimeOverhead,
                            double *totalTimeOverhead, PQ_t pq, process_t *minRemainTimeProcess_ptr);
 
-void HandleRemainProcOnOneCPU(unsigned int *globalTimer, unsigned int *totalDeltaTime, int *numRunnedProcesses,
+void HandleRemainProcOnOneCPU(unsigned long *globalTimer, unsigned int *totalDeltaTime, int *numRunnedProcesses,
                               double *maxTimeOverhead, double *totalTimeOverhead, PQ_t pq,
                               process_t *minRemainTimeProcess_ptr);
 
@@ -37,12 +37,12 @@ void InitMemCpuForMultiProc(int numMainProcess, int numCPU, HashTable *parallelP
                             process_t ***cpuMinProcess_ptr_list);
 
 void HandleArrivingProcOnMultiCpu(input_node_ptr input_list_head, int numCPU, HashTable *parallelProcTable,
-                                  input_node_ptr input_process_ptr, unsigned int *globalTimer,
+                                  input_node_ptr input_process_ptr, unsigned long *globalTimer,
                                   unsigned int *totalDeltaTime, int *numRunnedProcesses, double *maxTimeOverhead,
                                   double *totalTimeOverhead, int procCounter, PQ_t *cpuPQList,
                                   process_t **cpuMinProcess_ptr_list);
 
-void HandleRemainProcOnMultiCPU(int numCPU, HashTable *parallelProcTable, unsigned int *globalTimer,
+void HandleRemainProcOnMultiCPU(int numCPU, HashTable *parallelProcTable, unsigned long *globalTimer,
                                 unsigned int *totalDeltaTime, int *numRunnedProcesses, double *maxTimeOverhead,
                                 double *totalTimeOverhead, int procCounter, PQ_t *cpuPQList,
                                 process_t **cpuMinProcess_ptr_list);
